further refactoring: moved more classes to deprecated, deleted cells folder
