+++ src/main/java/deprecated/classes/FlagCell.java
@@ -1,8 +1,8 @@
-package cells;
+package deprecated.classes;
 
 import interfaces.ICell;
 
-public class BlankCell implements ICell {
+public class FlagCell implements ICell {
 
     public String toString(){
         return "BlankCell";
