+++ src/main/java/deprecated/classes/HoleCell.java
@@ -1,8 +1,8 @@
-package cells;
+package deprecated.classes;
 
 import interfaces.ICell;
 
-public class FlagCell implements ICell {
+public class HoleCell implements ICell {
 
     public String toString(){
         return "BlankCell";
